+++ resources/lang/de.php
@@ -204,6 +204,7 @@ return [
     'action_delete_tenant' => 'Mandant löschen',
     'action_sync_tenants' => 'Fehlende Mandanten suchen',
     'action_renew_ssl' => 'SSL erneuern',
+    'action_upgrade_docker' => 'Docker aktualisieren',
     'action_send_welcome_mail' => 'Willkommensmail senden',
     'help_admin_pass' => 'Definiert das Admin-Passwort des neuen Mandanten. Bleibt das Feld leer, ' .
         'wird ein zufälliges Passwort erzeugt',