+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 
 ### Feat
 
+- *(admin)* Add button to upgrade tenant Docker containers
 - Skip schema init when db ready
 - Support soft hyphen placeholders
 - Add dynamic congratulations for team PDFs